Fix camera crash on bad pmem allocations.
ADSP allocation table indicated conflicts due to believing a pmem block
was larger than its actual data size.

This is because PMEM provided the block size back to adsp and not the
allocated request size.

A better fix is to ask userspace to provide the requested length, however
this breaks current APIs such as libcamera; Thus to allow the system to
behave as expected store the original requested size, and apply it in
ADSP's logic. [for allocated regions only]

Previous versions of the kernel appeared to have hidden this with more
unique VMA mapping, however I believe this is a very long standing bug.